+++ hash-addressed/library/HashAddressed/HashFunction.hs
@@ -1,12 +1,25 @@
-module HashAddressed.HashFunction where
+module HashAddressed.HashFunction
+  (
+    {- * Type -} HashFunction (..),
+    {- * Examples -} sha256,
+  )
+  where
 
 import Essentials
 
-{-| Hash function supported by the @hash-addressed@ library
+import Fold.Pure (Fold (..))
+import System.IO (FilePath)
 
-    Currently only SHA-256 is supported, but others
-    may be added in the future.
--}
-data HashFunction =
-    SHA_256 -- ^ SHA-256
-  deriving stock (Eq, Ord)
+import qualified Crypto.Hash.SHA256 as Hash
+import qualified Data.ByteString as Strict
+import qualified Data.ByteString.Base16 as Base16
+import qualified Data.ByteString.Char8 as Strict.ByteString.Char8
+
+newtype HashFunction = HashFunction (Fold Strict.ByteString FilePath)
+
+sha256 :: HashFunction
+sha256 = HashFunction Fold
+  { initial = Hash.init
+  , step = Hash.update
+  , extract = Strict.ByteString.Char8.unpack . Base16.encode . Hash.finalize
+  }